@@ -1,8 +1,9 @@
 apply plugin: 'com.android.application'
+apply plugin: 'kotlin-android'
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion "28.0.3"
+    compileSdkVersion 29
+    buildToolsVersion "29.0.2"
 
     compileOptions {
         sourceCompatibility '1.8'
@@ -11,12 +12,13 @@ android {
 
     defaultConfig {
         applicationId "io.scalaproject.androidminer"
-        minSdkVersion 17
-        targetSdkVersion 28
-        versionCode 3
-        versionName "1.2.0"
+        minSdkVersion 19
+        targetSdkVersion 28 // keep version 28 to avoid an error 13 when running exe in assets folder (sdk 29 gives the error)
+        versionCode 10
+        versionName "2.0.0"
         buildConfigField "long", "BUILD_TIME", System.currentTimeMillis() + "L"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        vectorDrawables.useSupportLibrary = true
     }
 
     buildTypes {
@@ -28,12 +30,17 @@ android {
 }
 
 dependencies {
-    implementation 'com.android.support.constraint:constraint-layout:2.0.0-beta4'
-    implementation 'com.android.support:design:28.0.0'
-    implementation 'com.google.android.gms:play-services-vision:9.4.0+'
+    //noinspection GradleDependency
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3' // Keep the version 1.1.3
+    //noinspection GradleDependency
+    implementation 'com.google.android.material:material:1.1.0' // Keep the version 1.1.0
     implementation 'com.github.KingsMentor:MobileVisionBarcodeScanner:2.0.0'
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support:support-v4:28.0.0'
+    //noinspection GradleDependency
+    implementation 'androidx.appcompat:appcompat:1.1.0' // Keep the version 1.1.0
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation 'org.ocpsoft.prettytime:prettytime:4.0.2.Final'
     implementation 'com.android.volley:volley:1.1.1'
+    implementation 'com.github.anastr:speedviewlib:1.5.2'
+    implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk7:1.4.10'
+    implementation 'com.github.PhilJay:MPAndroidChart:v3.1.0'
 }
