@@ -2,31 +2,35 @@
 //
 // Please see the included LICENSE file for more information.
 //
-// Copyright (c) 2020, Scala
+// Copyright (c) 2021 Scala
 //
 // Please see the included LICENSE file for more information.
 
 package io.scalaproject.androidminer;
 
+import android.app.Dialog;
 import android.content.Intent;
+import android.graphics.Color;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 import android.text.Html;
+import android.text.SpannableString;
+import android.text.Spanned;
+import android.text.TextPaint;
 import android.text.method.LinkMovementMethod;
+import android.text.style.ClickableSpan;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
+import android.widget.ImageView;
+import android.widget.LinearLayout;
 import android.widget.TextView;
-import android.widget.Toast;
 
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
-import java.util.Calendar;
-import java.util.Locale;
-import java.util.Map;
+import io.scalaproject.androidminer.dialogs.CreditsFragment;
+import io.scalaproject.androidminer.dialogs.DonationsFragment;
 
 public class AboutFragment extends Fragment {
     @Nullable
@@ -34,27 +38,62 @@ public class AboutFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         View view = inflater.inflate(R.layout.fragment_about, container, false);
 
-        TextView tvBuild, tvScala, tvMine2gether, tvMonerominer, tvFontAwesome;
+        TextView tvBuild = view.findViewById(R.id.build);
+        tvBuild.setText(BuildConfig.VERSION_NAME + " (" + Utils.getBuildTime() + ")");
 
-        tvBuild = view.findViewById(R.id.build);
-
-        tvScala = view.findViewById(R.id.ScalaURL);
-        tvMine2gether = view.findViewById(R.id.Mine2getherURL);
-        tvMonerominer = view.findViewById(R.id.MoneroMinerURL);
-        tvFontAwesome = view.findViewById(R.id.FontAwesomeURL);
-
-        Button btnGitHub = view.findViewById(R.id.btnGitHub);
-        btnGitHub.setOnClickListener(new View.OnClickListener() {
+        TextView tvScala = view.findViewById(R.id.tvScala);
+        tvScala.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Uri uri = Uri.parse(getResources().getString(R.string.githubLink));
-                Intent intent = new Intent(Intent.ACTION_VIEW, uri);
+                Uri uri = Uri.parse(getString(R.string.ScalaLinkClean));
+                startActivity(new Intent(Intent.ACTION_VIEW, uri));
+            }
+        });
+
+        LinearLayout llGetSupport = view.findViewById(R.id.llSupport);
+        llGetSupport.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Intent intent = new Intent(getActivity(), SupportActivity.class);
                 startActivity(intent);
             }
         });
 
-        Button btnDiscord = view.findViewById(R.id.btnDiscord);
-        btnDiscord.setOnClickListener(new View.OnClickListener() {
+        LinearLayout llCredits = view.findViewById(R.id.llCredits);
+        llCredits.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                onShowCredits();
+            }
+        });
+
+        LinearLayout llShare = view.findViewById(R.id.llShare);
+        llShare.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                onShareApp();
+            }
+        });
+
+        LinearLayout llDonations = view.findViewById(R.id.llDonations);
+        llDonations.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                onShowDonations();
+            }
+        });
+
+        LinearLayout llPrivacyPolicy = view.findViewById(R.id.llPrivacyPolicy);
+        llPrivacyPolicy.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Uri uri = Uri.parse(getResources().getString(R.string.privacyPolicyLink));
+                startActivity(new Intent(Intent.ACTION_VIEW, uri));
+            }
+        });
+
+        ImageView ivDiscord = view.findViewById(R.id.ivDiscord);
+        ivDiscord.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
                 Uri uri = Uri.parse(getResources().getString(R.string.discordLink));
@@ -63,28 +102,8 @@ public class AboutFragment extends Fragment {
             }
         });
 
-        Button btnMedium = view.findViewById(R.id.btnMedium);
-        btnMedium.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Uri uri = Uri.parse(getResources().getString(R.string.mediumLink));
-                Intent intent = new Intent(Intent.ACTION_VIEW, uri);
-                startActivity(intent);
-            }
-        });
-
-        Button btnTwitter = view.findViewById(R.id.btnTwitter);
-        btnTwitter.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Uri uri = Uri.parse(getResources().getString(R.string.twitterLink));
-                Intent intent = new Intent(Intent.ACTION_VIEW, uri);
-                startActivity(intent);
-            }
-        });
-
-        Button btnTelegram = view.findViewById(R.id.btnTelegram);
-        btnTelegram.setOnClickListener(new View.OnClickListener() {
+        ImageView ivTelegram = view.findViewById(R.id.ivTelegram);
+        ivTelegram.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
                 Uri uri = Uri.parse(getResources().getString(R.string.telegramLink));
@@ -93,96 +112,100 @@ public class AboutFragment extends Fragment {
             }
         });
 
-        Button btnDonationAddressesHelp = view.findViewById(R.id.btnDonationsHelp);
-        btnDonationAddressesHelp.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                // inflate the layout of the popup window
-                View popupView = inflater.inflate(R.layout.helper_donation_addresses, null);
-                Utils.showPopup(v, inflater, popupView);
-            }
-        });
-
-        Button btnDonateBTC = view.findViewById(R.id.btnDonateBTC);
-        btnDonateBTC.setOnClickListener(new View.OnClickListener() {
+        ImageView ivMedium = view.findViewById(R.id.ivMedium);
+        ivMedium.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Utils.copyToClipboard("Scala BTC Donation Address", Utils.SCALA_BTC_ADDRESS);
-                Toast.makeText(getContext(), getResources().getString(R.string.donationadressbtc_copied), Toast.LENGTH_SHORT).show();
+                Uri uri = Uri.parse(getResources().getString(R.string.mediumLink));
+                Intent intent = new Intent(Intent.ACTION_VIEW, uri);
+                startActivity(intent);
             }
         });
 
-        Button btnDonateETH = view.findViewById(R.id.btnDonateETH);
-        btnDonateETH.setOnClickListener(new View.OnClickListener() {
+        ImageView ivTwitter = view.findViewById(R.id.ivTwitter);
+        ivTwitter.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Utils.copyToClipboard("Scala ETH Donation Address", Utils.SCALA_ETH_ADDRESS);
-                Toast.makeText(getContext(), getResources().getString(R.string.donationadresseth_copied), Toast.LENGTH_SHORT).show();
+                Uri uri = Uri.parse(getResources().getString(R.string.twitterLink));
+                Intent intent = new Intent(Intent.ACTION_VIEW, uri);
+                startActivity(intent);
             }
         });
 
-        Button btnDonateXLA = view.findViewById(R.id.btnDonateXLA);
-        btnDonateXLA.setOnClickListener(new View.OnClickListener() {
+        ImageView ivReddit = view.findViewById(R.id.ivReddit);
+        ivReddit.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Utils.copyToClipboard("Scala XLA Donation Address", Utils.SCALA_XLA_ADDRESS);
-                Toast.makeText(getContext(), getResources().getString(R.string.donationadressxla_copied), Toast.LENGTH_SHORT).show();
+                Uri uri = Uri.parse(getResources().getString(R.string.twitterLink));
+                Intent intent = new Intent(Intent.ACTION_VIEW, uri);
+                startActivity(intent);
             }
         });
 
-        StringBuilder cpuinfo = new StringBuilder(Config.read("CPUINFO").trim());
-        if(cpuinfo.length() == 0) {
-            try {
-                Map<String, String> m = Tools.getCPUInfo();
+        String sDisclaimerText = getResources().getString(R.string.disclaimer_agreement);
+        String sDiclaimer = getResources().getString(R.string.disclaimer);
 
-                cpuinfo = new StringBuilder("ABI: " + Tools.getABI() + "\n");
-                for (Map.Entry<String, String> pair : m.entrySet()) {
-                    cpuinfo.append(pair.getKey()).append(": ").append(pair.getValue()).append("\n");
-                }
-            } catch (Exception e) {
-                cpuinfo = new StringBuilder();
-            }
-
-            Config.write("CPUINFO", cpuinfo.toString().trim());
-        }
-
-        // Convert build time to readable date
-        Calendar calendar = Calendar.getInstance();
-        calendar.setTimeInMillis(BuildConfig.BUILD_TIME);
-        String build_time = DateFormat.getDateInstance(DateFormat.LONG).format(calendar.getTime());
-
-        tvBuild.setText(BuildConfig.VERSION_NAME + " (" + build_time + ")");
-
-        tvScala.setText(Html.fromHtml(getString(R.string.ScalaLink)));
-        tvScala.setMovementMethod(LinkMovementMethod.getInstance());
-
-        tvMine2gether.setText(Html.fromHtml(getString(R.string.Mine2getherLink)));
-        tvMine2gether.setMovementMethod(LinkMovementMethod.getInstance());
-
-        tvMonerominer.setText(Html.fromHtml(getString(R.string.MoneroMinerLink)));
-        tvMonerominer.setMovementMethod(LinkMovementMethod.getInstance());
-
-        tvFontAwesome.setText(Html.fromHtml(getString(R.string.FontAwesomeLink)));
-        tvFontAwesome.setMovementMethod(LinkMovementMethod.getInstance());
-
-        SimpleDateFormat formatter = new SimpleDateFormat("dd/MM/yyyy hh:mm a", Locale.ENGLISH);
-        String build_time_debug = formatter.format(calendar.getTime());
-
-        String sDebugInfo = "Version Code: " + BuildConfig.VERSION_CODE + "\n" +
-                "Version Name: " + BuildConfig.VERSION_NAME + "\n" +
-                "Build Time: " + build_time_debug + "\n\n" +
-                "Device Name: " + Tools.getDeviceName() + "\n" +
-                "CPU Info: " + cpuinfo;
-
-        Button btnDebugInfo = view.findViewById(R.id.btnDebugInfo);
-        btnDebugInfo.setOnClickListener(new View.OnClickListener() {
+        SpannableString ss = new SpannableString(sDisclaimerText);
+        ClickableSpan clickableSpan = new ClickableSpan() {
             @Override
-            public void onClick(View view) {
-                Utils.copyToClipboard("Scala Miner Debug Info", sDebugInfo);
-                Toast.makeText(getContext(), getResources().getString(R.string.debuginfo_copied), Toast.LENGTH_SHORT).show();
+            public void onClick(View textView) {
+                showDisclaimer();
             }
-        });
+
+            @Override
+            public void updateDrawState(TextPaint ds) {
+                super.updateDrawState(ds);
+                ds.setUnderlineText(false);
+            }
+        };
+
+        int iStart = sDisclaimerText.indexOf(sDiclaimer);
+        int iEnd = iStart + sDiclaimer.length();
+        ss.setSpan(clickableSpan, iStart, iEnd, Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
+
+        TextView tvDisclaimer = view.findViewById(R.id.disclaimer);
+        tvDisclaimer.setText(ss);
+        tvDisclaimer.setMovementMethod(LinkMovementMethod.getInstance());
+        tvDisclaimer.setLinkTextColor(getResources().getColor(R.color.c_blue));
+        tvDisclaimer.setHighlightColor(Color.TRANSPARENT);
 
         return view;
+    }
+
+    private void onShowCredits() {
+        CreditsFragment.display(getActivity().getSupportFragmentManager());
+    }
+
+    private void onShowDonations() {
+        DonationsFragment.display(getActivity().getSupportFragmentManager());
+    }
+
+    private void onShareApp() {
+        Intent sharingIntent = new Intent(android.content.Intent.ACTION_SEND);
+        sharingIntent.setType("text/plain");
+        String shareBody = getResources().getString(R.string.app_share);
+        sharingIntent.putExtra(android.content.Intent.EXTRA_SUBJECT, getResources().getString(R.string.share_title));
+        sharingIntent.putExtra(android.content.Intent.EXTRA_TEXT, shareBody);
+
+        startActivity(Intent.createChooser(sharingIntent, getResources().getString(R.string.share_via)));
+    }
+
+    private void showDisclaimer() {
+        final Dialog dialog = new Dialog(getContext());
+        dialog.setContentView(R.layout.disclaimer);
+        dialog.setTitle("Disclaimer");
+        dialog.setCancelable(false);
+
+        Button btnOK = dialog.findViewById(R.id.btnAgree);
+
+        btnOK.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Config.write("disclaimer_agreed", "1");
+                dialog.dismiss();
+            }
+        });
+
+        dialog.show();
     }
 }
