@@ -20,12 +20,13 @@
 //
 // Please see the included LICENSE file for more information.
 //
-// Copyright (c) 2020, Scala
+// Copyright (c) 2021 Scala
 //
 // Please see the included LICENSE file for more information.
 
 package io.scalaproject.androidminer;
 
+import android.app.NotificationManager;
 import android.app.Service;
 import android.content.Context;
 import android.content.Intent;
@@ -37,6 +38,13 @@ import android.util.Log;
 import android.widget.Toast;
 import android.os.AsyncTask;
 
+import com.android.volley.RequestQueue;
+import com.android.volley.toolbox.JsonObjectRequest;
+import com.android.volley.toolbox.RequestFuture;
+import com.android.volley.toolbox.Volley;
+
+import org.json.JSONObject;
+
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.File;
@@ -45,11 +53,11 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
-import java.net.Inet4Address;
-import java.net.Inet6Address;
-import java.net.InetAddress;
-import java.net.UnknownHostException;
 import java.util.Arrays;
+import java.util.Objects;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 
 import io.scalaproject.androidminer.api.PoolItem;
 import io.scalaproject.androidminer.api.ProviderManager;
@@ -67,23 +75,39 @@ public class MiningService extends Service {
     private ProcessMonitor procMon;
     private PowerManager.WakeLock wl;
     private int accepted = 0;
-    private String speed = "0";
+    private int difficulty = 0;
+    private int connection = 0;
+    private float speed = 0.0f;
+    private float max = 0.0f;
     private String lastAssetPath;
     private String lastOutput = "";
+    private static RequestQueue reqQueue;
+
+    private static String API_IP = "https://json.geoiplookup.io/";
 
     @Override
-    public void onCreate() {
+    public void onCreate()
+    {
         super.onCreate();
         privatePath = getFilesDir().getAbsolutePath();
         Tools.deleteDirectoryContents(new File(privatePath));
+
+        reqQueue = Volley.newRequestQueue(this);
+    }
+
+    @Override
+    public void onTaskRemoved(Intent rootIntent) {
+        super.onTaskRemoved(rootIntent);
+
+        NotificationManager notificationManager = (NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);
+        notificationManager.cancelAll();
     }
 
     private MiningServiceStateListener listener = null;
 
     public interface MiningServiceStateListener {
-        public void onStateChange(Boolean state);
-
-        public void onStatusChange(String status, String speed, Integer accepted);
+        void onStateChange(Boolean state);
+        void onStatusChange(String status, float speed, float max, Integer accepted, Integer difficulty, Integer connection);
     }
 
     public void setMiningServiceStateListener(MiningServiceStateListener listener) {
@@ -98,8 +122,8 @@ public class MiningService extends Service {
         if (listener != null) listener.onStateChange(state);
     }
 
-    private void raiseMiningServiceStatusChange(String status, String speed, Integer accepted) {
-        if (listener != null) listener.onStatusChange(status, speed, accepted);
+    private void raiseMiningServiceStatusChange(String status, float speed, float max, Integer accepted, Integer difficulty, Integer connection) {
+        if (listener != null) listener.onStatusChange(status, speed, max, accepted, difficulty, connection);
     }
 
     public Boolean getMiningServiceState() {
@@ -107,7 +131,6 @@ public class MiningService extends Service {
     }
 
     private void copyMinerFiles() {
-
         String abi = Tools.getABI();
         String assetPath = "";
         String libraryPath = "";
@@ -146,7 +169,6 @@ public class MiningService extends Service {
     }
 
     private static String createCpuConfig(int cores, int threads, int intensity) {
-
         String cpuConfig = "";
 
         for (int i = 0; i < cores; i++) {
@@ -154,7 +176,7 @@ public class MiningService extends Service {
                 if (!cpuConfig.equals("")) {
                     cpuConfig += ",";
                 }
-                cpuConfig += "[" + Integer.toString(intensity) + "," + Integer.toString(i) + "]";
+                cpuConfig += "[" + intensity + "," + i + "]";
             }
         }
 
@@ -218,25 +240,23 @@ public class MiningService extends Service {
     }
 
     public static String getIpByHost(PoolItem pi) {
-        String hostName = pi.getPool() + ":" + pi.getPort();
+        String hostIP = "";
+
+        RequestFuture<JSONObject> future = RequestFuture.newFuture();
+        JsonObjectRequest request = new JsonObjectRequest(API_IP + pi.getPool(), new JSONObject(), future, future);
+        reqQueue.add(request);
 
         try {
-            Log.i(LOG_TAG, hostName);
-            InetAddress inetAddress = Inet4Address.getByName(hostName);
-
-            if(!inetAddress.isLoopbackAddress()) {
-                if (inetAddress instanceof Inet6Address) {
-                    return pi.getPoolIP(); // workaround for IPv6
-                } else if (inetAddress instanceof Inet4Address) {
-                    return inetAddress.getHostAddress().toString();
-                }
-            }
-        } catch (UnknownHostException e) {
-            Log.i(LOG_TAG, e.toString());
-            return hostName;
+            JSONObject response = future.get(5, TimeUnit.SECONDS); // Sync call
+            hostIP = response.optString("ip");
+        } catch (InterruptedException | ExecutionException | TimeoutException e) {
+            e.printStackTrace();
         }
 
-        return hostName;
+        if(hostIP.isEmpty())
+            hostIP = pi.getPoolIP();
+
+        return hostIP + ":" + pi.getPort();
     }
 
     public void startMining(MiningConfig config) {
@@ -262,6 +282,7 @@ public class MiningService extends Service {
                 this.config.pool = getPoolHost();
                 return "success";
             } catch (Exception e) {
+                e.printStackTrace();
             }
             return null;
         }
@@ -273,7 +294,6 @@ public class MiningService extends Service {
     }
 
     public void startMiningProcess(MiningConfig config) {
-
         Log.i(LOG_TAG, "starting...");
 
         if (process != null) {
@@ -306,7 +326,10 @@ public class MiningService extends Service {
             pb.redirectErrorStream();
 
             accepted = 0;
-            speed = "n/a";
+            difficulty = 0;
+            connection = 0;
+            speed = -1.0f;
+            max = -1.0f;
             lastOutput = "";
 
             process = pb.start();
@@ -326,12 +349,12 @@ public class MiningService extends Service {
 
         } catch (Exception e) {
             Log.e(LOG_TAG, "exception:", e);
-            Toast.makeText(this, e.getLocalizedMessage(), Toast.LENGTH_SHORT).show();
+            Utils.showToast(this, e.getLocalizedMessage(), Toast.LENGTH_SHORT);
             process = null;
         }
     }
 
-    public String getSpeed() {
+    public float getSpeed() {
         return speed;
     }
 
@@ -381,43 +404,74 @@ public class MiningService extends Service {
     }
 
     private class OutputReaderThread extends Thread {
-
         private InputStream inputStream;
-        private BufferedReader reader;
         private StringBuilder output = new StringBuilder();
 
         OutputReaderThread(InputStream inputStream, String miner) {
-
             this.inputStream = inputStream;
         }
 
         private void processLogLine(String line) {
-            output.append(line + System.getProperty("line.separator"));
+            output.append(line).append(System.getProperty("line.separator"));
 
             String lineCompare = line.toLowerCase();
             if (lineCompare.contains("accepted")) {
                 accepted++;
-            } else if (lineCompare.contains("speed")) {
+
+                if(lineCompare.contains("diff")) {
+                    int i = lineCompare.indexOf("diff ") + "diff ".length();
+                    int imax = lineCompare.indexOf(" ", i);
+                    String diff = lineCompare.substring(i, imax).trim();
+                    difficulty = Integer.parseInt(lineCompare.substring(i, imax).trim());;
+                }
+
+                if(lineCompare.contains("ms)")) {
+                    int i = lineCompare.indexOf("(", lineCompare.length() - 10) + 1;
+                    int imax = lineCompare.indexOf("ms)");
+                    String conn = lineCompare.substring(i, imax).trim();
+                    connection = Integer.parseInt(lineCompare.substring(i, imax).trim());;
+                }
+
+            } // For some reason some devices display "miner" instead of "speed"
+            else if (lineCompare.contains("speed")) {
                 String[] split = TextUtils.split(line, " ");
-                speed = split[4];
-                if (speed.equals("n/a")) {
-                    speed = split[5];
-                    if (speed.equals("n/a")) {
-                        speed = split[6];
+                String tmpSpeed = lineCompare.contains("miner") ? split[9] : split[4];
+                if (tmpSpeed.equals("n/a")) {
+                    tmpSpeed = split[5];
+                    if (tmpSpeed.equals("n/a")) {
+                        tmpSpeed = split[6];
+                    }
+                }
+
+                try {
+                    speed = Float.parseFloat(tmpSpeed.trim());
+                } catch (NumberFormatException e) {
+                    // Ignore
+                }
+
+                if (lineCompare.contains("max")) {
+                    int i = lineCompare.indexOf("max ") + "max ".length();
+                    int imax = lineCompare.indexOf(" ", i);
+                    String tmpMax = lineCompare.substring(i, imax).trim();
+
+                    try {
+                        max = Float.parseFloat(tmpMax);
+                    } catch (NumberFormatException e) {
+                        // Ignore
                     }
                 }
             }
 
             if (output.length() > Config.logMaxLength) {
-                output.delete(0, output.indexOf(System.getProperty("line.separator"), Config.logPruneLength) + 1);
+                output.delete(0, output.indexOf(Objects.requireNonNull(System.getProperty("line.separator")), Config.logPruneLength) + 1);
             }
 
-            raiseMiningServiceStatusChange(line, speed, accepted);
+            raiseMiningServiceStatusChange(line, speed, max, accepted, difficulty, connection);
         }
 
         public void run() {
             try {
-                reader = new BufferedReader(new InputStreamReader(inputStream));
+                BufferedReader reader = new BufferedReader(new InputStreamReader(inputStream));
                 String line;
                 while ((line = reader.readLine()) != null) {
 
