@@ -1,25 +1,33 @@
-// Copyright (c) 2020, Scala
+// Copyright (c) 2021 Scala
 //
 // Please see the included LICENSE file for more information.
 
 package io.scalaproject.androidminer;
 
-import android.app.Activity;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.os.Handler;
 
-public class SplashActivity extends Activity {
+import com.google.android.gms.common.GooglePlayServicesNotAvailableException;
+import com.google.android.gms.common.GooglePlayServicesRepairableException;
+import com.google.android.gms.security.ProviderInstaller;
+
+import java.security.KeyManagementException;
+import java.security.NoSuchAlgorithmException;
+
+import javax.net.ssl.SSLContext;
+
+public class SplashActivity extends BaseActivity {
     @Override
     public void onCreate(Bundle savedInstanceState) {
         SharedPreferences preferences = getSharedPreferences(getPackageName() + "_preferences", MODE_PRIVATE);
         Config.initialize(preferences);
 
-        String configversion = Config.read("config_version");
+        String configversion = Config.read(Config.CONFIG_KEY_CONFIG_VERSION);
         if(!configversion.equals(Config.version)) {
             Config.clear();
-            Config.write("config_version", Config.version);
+            Config.write(Config.CONFIG_KEY_CONFIG_VERSION, Config.version);
         }
 
         super.onCreate(savedInstanceState);
@@ -32,13 +40,24 @@ public class SplashActivity extends Activity {
 
         setContentView(R.layout.splashscreen);
 
+        try {
+            ProviderInstaller.installIfNeeded(getApplicationContext());
+            SSLContext sslContext;
+            sslContext = SSLContext.getInstance("TLSv1.2");
+            sslContext.init(null, null, null);
+            sslContext.createSSLEngine();
+        } catch (GooglePlayServicesRepairableException | GooglePlayServicesNotAvailableException
+                | NoSuchAlgorithmException | KeyManagementException e) {
+            e.printStackTrace();
+        }
+
         int millisecondsDelay = 2000;
         new Handler().postDelayed(new Runnable() {
             public void run() {
-                String hide_landing_page = Config.read("hide_landing_page");
+                String hide_setup_wizard = Config.read("hide_setup_wizard");
 
-                if (hide_landing_page.equals("")) {
-                    startActivity(new Intent(SplashActivity.this, LandingPageActivity.class));
+                if (hide_setup_wizard.equals("")) {
+                    startActivity(new Intent(SplashActivity.this, WizardHomeActivity.class));
                 } else {
                     startActivity(new Intent(SplashActivity.this, MainActivity.class));
                 }
