@@ -2,27 +2,25 @@
 //
 // Please see the included LICENSE file for more information.
 //
-// Copyright (c) 2020, Scala
+// Copyright (c) 2021 Scala
 //
 // Please see the included LICENSE file for more information.
 
 package io.scalaproject.androidminer;
 
 import android.content.Intent;
-import android.graphics.drawable.Drawable;
-import android.net.Uri;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v4.graphics.drawable.DrawableCompat;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Button;
+import android.widget.ImageView;
+import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
+
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 
 import io.scalaproject.androidminer.api.ProviderData;
 import io.scalaproject.androidminer.api.PoolItem;
@@ -30,18 +28,16 @@ import io.scalaproject.androidminer.api.IProviderListener;
 import io.scalaproject.androidminer.api.ProviderManager;
 
 public class StatsFragment extends Fragment {
+    protected static IProviderListener statsListener;
 
-    private static final String LOG_TAG = "MiningSvc";
+    public static ProviderData poolData = null;
 
-    private Button bStatCheckOnline;
+    static View view = null;
 
-    protected IProviderListener statsListener;
     @Nullable
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_stats, container, false);
-
-        bStatCheckOnline = view.findViewById(R.id.checkstatsonline);
+        view = inflater.inflate(R.layout.fragment_stats, container, false);
 
         statsListener = new IProviderListener() {
             public void onStatsChange(ProviderData d) {
@@ -54,28 +50,49 @@ public class StatsFragment extends Fragment {
             }
         };
 
+        LinearLayout llPayments = view.findViewById(R.id.llPayments);
+        llPayments.setOnClickListener(new View.OnClickListener() {
+            public void onClick(View v) {
+                onShowPayments();
+            }
+        });
+
         ProviderManager.request.setListener(statsListener).start();
         ProviderManager.afterSave();
+
         updateFields(ProviderManager.data, view);
 
         return view;
     }
 
-    private void updateFields(ProviderData d, View view) {
+    public static void updateStatsListener() {
+        ProviderManager.fetchStats();
+        updateFields(ProviderManager.data, view);
+    }
+
+    private static void updateFields(ProviderData d, View view) {
+        poolData = d;
+
         if(view == null || view.getContext() == null)
             return;
 
         if(d.isNew) {
-            enableOnlineStats(false);
             return;
         }
 
-        PoolItem pm = ProviderManager.getSelectedPool();
+        PoolItem pi = ProviderManager.getSelectedPool();
+
+        ImageView ivShowPayments = view.findViewById(R.id.ivShowPayments);
+        ivShowPayments.setVisibility(pi.getPoolType() == 0 ? View.GONE : View.VISIBLE);
 
         // Network
 
+        String[] n = d.network.hashrate.split(" ");
         TextView tvNetworkHashrate = view.findViewById(R.id.hashratenetwork);
-        tvNetworkHashrate.setText(d.network.hashrate.isEmpty() ? "n/a" : d.network.hashrate);
+        tvNetworkHashrate.setText(n.length > 0 ? n[0] : "n/a");
+
+        TextView tvNetworkHashrateUnit = view.findViewById(R.id.hashratenetwork_unit);
+        tvNetworkHashrateUnit.setText(n.length > 1 ? n[1] : "MH/s");
 
         TextView tvNetworkDifficulty = view.findViewById(R.id.difficultynetwork);
         tvNetworkDifficulty.setText(d.network.difficulty.isEmpty() ? "n/a" : d.network.difficulty);
@@ -84,7 +101,7 @@ public class StatsFragment extends Fragment {
         tvNetworkBlocks.setText(d.network.lastBlockTime.isEmpty() ? "n/a" : d.network.lastBlockTime);
 
         TextView tvNetworkHeight = view.findViewById(R.id.height);
-        tvNetworkHeight.setText(d.network.lastBlockHeight.isEmpty() ? "n/a" : d.network.lastBlockHeight);
+        tvNetworkHeight.setText(d.network.lastBlockHeight.isEmpty() ? "n/a" : String.format("%,d", Integer.parseInt(d.network.lastBlockHeight)));
 
         TextView tvNetworkRewards = view.findViewById(R.id.rewards);
         tvNetworkRewards.setText(d.network.lastRewardAmount.isEmpty() ? "n/a" : d.network.lastRewardAmount);
@@ -92,78 +109,78 @@ public class StatsFragment extends Fragment {
         // Pool
 
         TextView tvPoolURL = view.findViewById(R.id.poolurl);
-        tvPoolURL.setText(pm.getPool() == null ? "" : pm.getPool());
+        tvPoolURL.setText(pi.getPool() == null ? "" : pi.getPool());
 
+        String[] p = d.pool.hashrate.split(" ");
         TextView tvPoolHashrate = view.findViewById(R.id.hashratepool);
-        tvPoolHashrate.setText(d.pool.hashrate.isEmpty() ? "n/a" : d.pool.hashrate);
+        tvPoolHashrate.setText(p.length > 0 ? p[0] : "n/a");
 
-        TextView tvPoolDifficulty = view.findViewById(R.id.difficultypool);
-        tvPoolDifficulty.setText(d.pool.difficulty.isEmpty() ? "n/a" : d.pool.difficulty);
+        TextView tvPoolHashrateUnit = view.findViewById(R.id.hashratepool_unit);
+        tvPoolHashrateUnit.setText(p.length > 1 ? p[1] : "kH/s");
 
-        TextView tvPoolBlocks = view.findViewById(R.id.lastblockpool);
-        tvPoolBlocks.setText(d.pool.lastBlockTime.isEmpty() ? "n/a" : d.pool.lastBlockTime);
+        TextView tvPoolMiners = view.findViewById(R.id.miners);
+        tvPoolMiners.setText(String.format("%,d", Integer.parseInt(d.pool.miners)));
 
-        TextView tvPoolLastBlock = view.findViewById(R.id.blockspool);
-        tvPoolLastBlock.setText(d.pool.blocks.isEmpty() ? "n/a" : d.pool.blocks);
+        LinearLayout llPoolBlocks = view.findViewById(R.id.llBlocksPool);
+        llPoolBlocks.setVisibility(pi.getPoolType() == 2 || pi.getPoolType() == 0 ? View.GONE : View.VISIBLE);
+
+        TextView tvPoolLastBlock = view.findViewById(R.id.lastblockpool);
+        tvPoolLastBlock.setText(d.pool.lastBlockTime.isEmpty() ? "n/a" : d.pool.lastBlockTime);
+
+        TextView tvPoolLBlocks = view.findViewById(R.id.blockspool);
+        tvPoolLBlocks.setText(d.pool.blocks.isEmpty() ? "n/a" : String.format("%,d", Integer.parseInt(d.pool.blocks)));
 
         // Address
 
         String wallet = Config.read("address");
-        String prettyaddress = wallet.substring(0, 7) + "..." + wallet.substring(wallet.length() - 7);
+        String prettyaddress = "";
+        if(!wallet.isEmpty())
+            prettyaddress = wallet.substring(0, 7) + "..." + wallet.substring(wallet.length() - 7);
 
         TextView tvWalletAddress = view.findViewById(R.id.walletaddress);
         tvWalletAddress.setText(prettyaddress);
 
         String sHashrate = d.miner.hashrate;
-        if(sHashrate != null) {
-            sHashrate = sHashrate.replace("H", "").trim();
-            TextView tvAddressHashrate = view.findViewById(R.id.hashrateminer);
-            tvAddressHashrate.setText(sHashrate);
+        TextView tvAddressHashrate = view.findViewById(R.id.hashrateminer);
+        TextView tvAddressHashrateUnit = view.findViewById(R.id.hashrateminer_unit);
 
-            TextView tvAddressLastShare = view.findViewById(R.id.lastshareminer);
-            tvAddressLastShare.setText(d.miner.lastShare.isEmpty() ? "n/a" : d.miner.lastShare);
+        if(!sHashrate.isEmpty()) {
+            String[] a = sHashrate.split(" ");
 
-            TextView tvAddressBlocks = view.findViewById(R.id.blocksminedminer);
-            tvAddressBlocks.setText(d.miner.blocks.isEmpty() ? "n/a" : d.miner.blocks);
-
-            String sBalance = d.miner.balance.replace("XLA", "").trim();
-            TextView tvBalance = view.findViewById(R.id.balance);
-            tvBalance.setText(sBalance);
-
-            String sPaid = d.miner.paid.replace("XLA", "").trim();
-            TextView tvPaid = view.findViewById(R.id.paid);
-            tvPaid.setText(sPaid);
+            tvAddressHashrate.setText(a.length > 0 ? a[0] : "n/a");
+            tvAddressHashrateUnit.setText(a.length > 1 ? a[1] : "H/s");
+        } else {
+            tvAddressHashrate.setText(pi.getPoolType() == 0 ? "n/a" : "0");
+            tvAddressHashrateUnit.setText("H/s");
         }
 
-        enableOnlineStats(true);
+        TextView tvAddressLastShare = view.findViewById(R.id.lastshareminer);
+        tvAddressLastShare.setText(d.miner.lastShare.isEmpty() ? "n/a" : d.miner.lastShare);
 
-        String statsUrlWallet = pm.getStatsURL() + "?wallet=" + wallet;
-        bStatCheckOnline.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Uri uri = Uri.parse(statsUrlWallet);
-                Intent intent = new Intent(Intent.ACTION_VIEW, uri);
-                startActivity(intent);
-            }
-        });
+        TextView tvAddressSubmittedHash = view.findViewById(R.id.submittedhash);
+        tvAddressSubmittedHash.setText(pi.getPoolType() == 1 || pi.getPoolType() == 2 ? view.getResources().getString(R.string.submitted_shares) : view.getResources().getString(R.string.submitted_hashes));
+
+        TextView tvAddressBlocks = view.findViewById(R.id.blocksminedminer);
+        tvAddressBlocks.setText(d.miner.shares.isEmpty() ? "n/a" : String.format("%,d", Integer.parseInt(d.miner.shares)));
+
+        String sBalance = d.miner.balance.replace("XLA", "").trim();
+        TextView tvBalance = view.findViewById(R.id.balance);
+        tvBalance.setText(d.miner.balance.isEmpty() ? pi.getPoolType() == 0 ? "n/a" : Tools.getLongValueString(0.0) : sBalance);
+
+        String sPaid = d.miner.paid.replace("XLA", "").trim();
+        TextView tvPaid = view.findViewById(R.id.paid);
+        tvPaid.setText(d.miner.paid.isEmpty() ? pi.getPoolType() == 0 ? "n/a" : Tools.getLongValueString(0.0) : sPaid);
+        tvPaid.setTextSize(sPaid.length() > 6 ? 12 : 14);
+
+        TextView tvPaidUnit = view.findViewById(R.id.paid_unit);
+        tvPaidUnit.setTextSize(sPaid.length() > 6 ? 12 : 14);
     }
 
-    private void enableOnlineStats(boolean enable) {
-        Drawable buttonDrawable = bStatCheckOnline.getBackground();
-        buttonDrawable = DrawableCompat.wrap(buttonDrawable);
+    public void onShowPayments() {
+        PoolItem pi = ProviderManager.getSelectedPool();
 
-        bStatCheckOnline.setEnabled(enable);
-
-        if (enable) {
-            DrawableCompat.setTint(buttonDrawable, getResources().getColor(R.color.bg_lighter));
-            bStatCheckOnline.setBackground(buttonDrawable);
-            bStatCheckOnline.setTextColor(getResources().getColor(R.color.c_white));
-        }
-        else {
-            DrawableCompat.setTint(buttonDrawable, getResources().getColor(R.color.bg_black));
-            bStatCheckOnline.setBackground(buttonDrawable);
-            bStatCheckOnline.setTextColor(getResources().getColor(R.color.c_black));
-        }
+        if(pi.getPoolType() != 0)
+            startActivity(new Intent(getActivity(), PaymentsActivity.class));
     }
 
     public boolean checkValidState() {
@@ -171,26 +188,21 @@ public class StatsFragment extends Fragment {
             return false;
 
         if(Config.read("address").equals("")) {
-            Toast.makeText(getContext(),"Wallet address is empty.", Toast.LENGTH_LONG).show();
-            enableOnlineStats(false);
+            Utils.showToast(getContext(),"Wallet address is empty.", Toast.LENGTH_LONG);
             return false;
         }
 
         PoolItem pi = ProviderManager.getSelectedPool();
 
         if (!Config.read("init").equals("1") || pi == null) {
-            Toast.makeText(getContext(),"Start mining to view statistics.", Toast.LENGTH_LONG).show();
-            enableOnlineStats(false);
+            Utils.showToast(getContext(),"Start mining to view statistics.", Toast.LENGTH_LONG);
             return false;
         }
 
         if (pi.getPoolType() == 0) {
-            Toast.makeText(getContext(),"Statistics are not available for custom pools.", Toast.LENGTH_LONG).show();
-            enableOnlineStats(false);
+            Utils.showToast(getContext(),"Statistics are not available for custom pools.", Toast.LENGTH_LONG);
             return false;
         }
-
-        enableOnlineStats(true);
 
         return true;
     }
@@ -199,17 +211,5 @@ public class StatsFragment extends Fragment {
     public void onResume() {
         super.onResume();
         ProviderManager.request.setListener(statsListener).start();
-    }
-
-    @Override
-    public void onPause() {
-        super.onPause();
-        enableOnlineStats(false);
-    }
-
-    @Override
-    public void onStop() {
-        super.onStop();
-        enableOnlineStats(false);
     }
 }
