@@ -1,8 +1,12 @@
-// Copyright (c) 2020, Scala
+// Copyright (c) 2021 Scala
 //
 // Please see the included LICENSE file for more information.
 
 package io.scalaproject.androidminer.api;
+
+import java.util.ArrayList;
+import java.util.Date;
+import java.util.List;
 
 public final class ProviderData {
     public static class Network {
@@ -15,12 +19,12 @@ public final class ProviderData {
 
     public static class Pool {
         public String lastBlockHeight = "";
-        public String difficulty = "";
         public String lastRewardAmount = "";
         public String lastBlockTime = "";
         public String hashrate = "";
         public String blocks = "";
         public String minPayout = "";
+        public String miners = "";
         int type = -1;
     }
 
@@ -29,7 +33,8 @@ public final class ProviderData {
         public String balance = "";
         public String paid = "";
         public String lastShare = "";
-        public String blocks = "";
+        public String shares = "";
+        public List<Payment> payments = new ArrayList<>();
     }
 
     public static class Coin {
@@ -39,10 +44,17 @@ public final class ProviderData {
         public long denominationUnit = 100L;
     }
 
+    public static class Payment {
+        public float amount;
+        public float fee;
+        public String hash;
+        public String timestamp;
+    }
+
     final public Network network = new Network();
     final public Pool pool = new Pool();
     final public Coin coin = new Coin();
-    final public Miner miner = new Miner();
+    public Miner miner = new Miner();
 
     public boolean isNew = true;
 }
