@@ -1,11 +1,26 @@
-// Copyright (c) 2020, Scala
+// Copyright (c) 2021 Scala
 //
 // Please see the included LICENSE file for more information.
 
 package io.scalaproject.androidminer.api;
 
+import android.content.Context;
+import android.content.SharedPreferences;
+import android.graphics.Bitmap;
+
+import org.json.JSONArray;
+import org.json.JSONException;
+import org.json.JSONObject;
+
 import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Map;
+
 import io.scalaproject.androidminer.Config;
+import io.scalaproject.androidminer.R;
+import io.scalaproject.androidminer.SettingsFragment;
+import io.scalaproject.androidminer.Utils;
+import io.scalaproject.androidminer.network.Json;
 
 public final class ProviderManager {
 
@@ -15,48 +30,108 @@ public final class ProviderManager {
         mPools.add(poolItem);
     }
 
-    static public void add(String key, String pool,String port, int poolType, String poolUrl, String poolIP) {
-        mPools.add(new PoolItem(key, pool, port, poolType, poolUrl, poolIP));
+    static public void delete(PoolItem poolItem) {
+        if(mPools.contains(poolItem))
+            mPools.remove(poolItem);
     }
 
-    static public void add(String key, String pool, String port, int poolType, String poolUrl, String poolIP, String poolApi) {
-        mPools.add(new PoolItem(key, pool, port, poolType, poolUrl, poolIP, poolApi, "",""));
+    static public PoolItem add(String key, String pool,String port, int poolType, String poolUrl, String poolIP) {
+        PoolItem pi = new PoolItem(key, pool, port, poolType, poolUrl, poolIP);
+        mPools.add(pi);
+
+        return pi;
     }
 
-    static public PoolItem[] getPools() {
+    static public PoolItem add(String key, String pool, String port, int poolType, String poolUrl, String poolIP, String poolApi) {
+        PoolItem pi = new PoolItem(key, pool, port, poolType, poolUrl, poolIP, poolApi, "","");
+        mPools.add(pi);
+
+        return pi;
+    }
+
+    static public Bitmap getDefaultPoolIcon(Context context, PoolItem poolItem) {
+        if(poolItem != null && poolItem.isOfficial())
+            return Utils.getCroppedBitmap(Utils.getBitmap(context, R.mipmap.ic_logo_blue));
+
+        return Utils.getCroppedBitmap(Utils.getBitmap(context, R.drawable.ic_pool_default));
+    }
+
+    static public void loadPools(Context context) {
+        loadDefaultPools();
+
+        loadUserdefinedPools(context);
+
+        // Selected pool
+        boolean selectedFound = false;
+        String sp = SettingsFragment.selectedPoolTmp == null ? Config.read(Config.CONFIG_SELECTED_POOL).trim() : SettingsFragment.selectedPoolTmp.getKey();
+        for(int i = 0; i < mPools.size(); i++) {
+            PoolItem pi = mPools.get(i);
+
+            if(pi.getKey().equals(sp)) {
+                selectedFound = true;
+                pi.setIsSelected(true);
+            } else {
+                pi.setIsSelected(false);
+            }
+        }
+
+        if(!selectedFound && !mPools.isEmpty()) {
+            Config.write(Config.CONFIG_SELECTED_POOL, mPools.get(0).getKey().trim());
+            mPools.get(0).setIsSelected(true);
+        }
+    }
+
+    static public PoolItem[] getPools(Context context) {
+        loadPools(context);
+
+        Collections.sort(mPools, PoolItem.PoolComparator);
+
         return mPools.toArray(new PoolItem[mPools.size()]);
     }
 
-    static public PoolItem getPoolById(int idx) {
-        return mPools.get(idx);
-    }
-
-    static public PoolItem getPoolById(String idx) {
-        int index = Integer.parseInt(idx);
-
-        if (idx.equals("") || mPools.size() < index || mPools.size() == 0) {
-            return null;
-        }
-
-        return mPools.get(index);
+    static public PoolItem[] getAllPools() {
+        return mPools.toArray(new PoolItem[mPools.size()]);
     }
 
     static final public ProviderData data = new ProviderData();
 
     static public PoolItem getSelectedPool() {
         if(request.mPoolItem != null) {
-            return  request.mPoolItem;
+            return request.mPoolItem;
         }
 
-        String sp = Config.read("selected_pool");
-        if (sp.equals("")) {
-            return null;
+        // Selected pool
+        PoolItem selectedPool = null;
+        if(SettingsFragment.selectedPoolTmp != null) {
+            for(int i = 0; i < mPools.size(); i++) {
+                selectedPool = mPools.get(i);
+
+                if(selectedPool.getKey().equals(SettingsFragment.selectedPoolTmp.getKey())) {
+                    return selectedPool;
+                }
+            }
+        } else {
+            String sp = Config.read(Config.CONFIG_SELECTED_POOL).trim();
+            for (int i = 0; i < mPools.size(); i++) {
+                PoolItem pi = mPools.get(i);
+
+                if (pi.getKey().equals(sp)) {
+                    selectedPool = pi;
+                    pi.setIsSelected(true);
+                } else {
+                    pi.setIsSelected(false);
+                }
+            }
         }
 
-        PoolItem pi = getPoolById(sp);
+        if(!mPools.isEmpty() && selectedPool == null) {
+            selectedPool = mPools.get(0);
+            selectedPool.setIsSelected(true);
+        }
 
-        return pi;
+        return selectedPool;
     }
+
     static public void afterSave() {
         if(request.mPoolItem != null)  {
             return;
@@ -67,7 +142,7 @@ public final class ProviderManager {
             return;
         }
 
-        mPools.clear();
+        //mPools.clear();
         request.mPoolItem = pi;
         data.isNew = true;
         request.start();
@@ -75,52 +150,90 @@ public final class ProviderManager {
 
     static final public ProviderRequest request = new ProviderRequest();
 
-    static public void generate() {
+    static public void fetchStats() {
+        request.run();
+    }
+
+    static public void loadDefaultPools() {
         request.stop();
         request.mPoolItem = null;
         mPools.clear();
 
-        // User Defined
-        add("custom", "custom", "3333", 0, "", "");
+        if(!mPools.isEmpty())
+            return;
 
-        // Scala Official pool
-        add(
-                "Scala Project (Official Pool)",
-                "mine.scalaproject.io",
-                "3333",
-                1,
-                "https://pool.scalaproject.io",
-                "198.204.241.13"
-        );
+        //String lastFetched = Config.read("RepositoryLastFetched");
+        String lastFetched = "";
+                String jsonString = "";
+        long now = System.currentTimeMillis() / 1000L;
 
-        // Another Scala pool : Miner.Rocks
-        add(
-                "Miner.Rocks",
-                "stellite.miner.rocks",
-                "5003",
-                2,
-                "https://stellite.miner.rocks",
-                "54.38.232.67"
-        );
+        if(!lastFetched.isEmpty() && Long.parseLong(lastFetched) < now){
+            jsonString = Config.read("RepositoryJson");
+        }
 
-        // Another Scala pool : HeroMiners
-        add(
-                "HeroMiners",
-                "scala.herominers.com",
-                "10130",
-                2,
-                "https://scala.herominers.com",
-                "138.201.217.40"
-        );
+        if(jsonString.isEmpty()) {
+            String url = Config.githubAppJson;
+            jsonString  = Json.fetch(url);
+            Config.write("RepositoryJson", jsonString);
+            Config.write("RepositoryLastFetched", String.valueOf(now + 3600));//Cached time is 1 hour for now
+        }
 
-        // Another Scala pool : GNTL
-        add(
-                "GNTL",
-                "xla.pool.gntl.co.uk",
-                "8080",
-                1,
-                "https://xla.pool.gntl.co.uk",
-                "83.151.238.38"
-        );
+        try {
+            JSONObject data = new JSONObject(jsonString);
+            JSONArray pools = data.getJSONArray("pools");
+
+            for(int i = 0; i < pools.length(); i++) {
+                JSONObject pool = pools.getJSONObject(i);
+
+                PoolItem poolItem;
+
+                if(!pool.has("apiUrl")) {
+                    poolItem = add(pool.getString("key"), pool.getString("pool"), pool.getString("port"), pool.getInt("poolType"), pool.getString("poolUrl"), pool.getString("poolIp"));
+                } else {
+                    poolItem = add(pool.getString("key"), pool.getString("pool"), pool.getString("port"), pool.getInt("poolType"), pool.getString("poolUrl"), pool.getString("poolIp"), pool.getString("apiUrl"));
+                }
+
+                // Icon
+                if(pool.has("icon")) {
+                    String iconURL = pool.getString("icon");
+                    if (!iconURL.isEmpty()) {
+                        Bitmap icon = Utils.getBitmapFromURL(iconURL);
+                        if(icon != null)
+                            poolItem.setIcon(Utils.getCroppedBitmap(icon));
+                    }
+                }
+            }
+        } catch (JSONException e) {
+            e.printStackTrace();
+        }
+    }
+
+    static public void loadUserdefinedPools(Context context) {
+        Map<String, ?> pools = context.getSharedPreferences(Config.CONFIG_USERDEFINED_POOLS, Context.MODE_PRIVATE).getAll();
+        for (Map.Entry<String, ?> poolEntry : pools.entrySet()) {
+            if (poolEntry != null) { // just in case, ignore possible future errors
+                PoolItem pi = PoolItem.fromString((String) poolEntry.getValue());
+                if (pi != null) {
+                    pi.setUserDefined(true);
+                    add(pi);
+                }
+            }
+        }
+    }
+
+    static public void saveUserDefinedPools(Context context) {
+        SharedPreferences.Editor editor = context.getSharedPreferences(Config.CONFIG_USERDEFINED_POOLS, Context.MODE_PRIVATE).edit();
+        editor.clear();
+
+        for(int i = 0; i < mPools.size(); i++) {
+            PoolItem pi = mPools.get(i);
+
+            if(pi.isUserDefined()) { // just in case!
+                String poolString = pi.toString();
+                editor.putString(Integer.toString(i), poolString);
+            }
+        }
+
+        editor.apply();
     }
 }
