@@ -2,7 +2,7 @@
 //
 // Please see the included LICENSE file for more information.
 //
-// Copyright (c) 2020, Scala
+// Copyright (c) 2021 Scala
 //
 // Please see the included LICENSE file for more information.
 
@@ -10,17 +10,24 @@ package io.scalaproject.androidminer.api.providers;
 
 import android.util.Log;
 
+import com.android.volley.Request;
+import com.android.volley.toolbox.StringRequest;
+
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.ocpsoft.prettytime.PrettyTime;
 
 import java.util.Date;
 
+import io.scalaproject.androidminer.Utils;
+import io.scalaproject.androidminer.PoolActivity;
 import io.scalaproject.androidminer.api.ProviderData;
 import io.scalaproject.androidminer.network.Json;
 import io.scalaproject.androidminer.api.ProviderAbstract;
 import io.scalaproject.androidminer.api.PoolItem;
+import io.scalaproject.androidminer.widgets.PoolInfoAdapter;
 
+import static io.scalaproject.androidminer.Tools.getReadableDifficultyString;
 import static io.scalaproject.androidminer.Tools.getReadableHashRateString;
 import static io.scalaproject.androidminer.Tools.parseCurrency;
 import static io.scalaproject.androidminer.Tools.tryParseLong;
@@ -31,6 +38,32 @@ public class CryptonoteNodejsPool extends ProviderAbstract {
         super(pi);
     }
 
+    public StringRequest getStringRequest(PoolInfoAdapter poolsAdapter) {
+        String url = mPoolItem.getApiUrl().isEmpty() ?  mPoolItem.getPool() : mPoolItem.getApiUrl();
+        url += "/stats";
+
+        return new StringRequest(Request.Method.GET, url,
+                response -> {
+                    try {
+                        JSONObject obj = new JSONObject(response);
+                        JSONObject objConfigPool = obj.getJSONObject("pool");
+
+                        // Miners
+                        mPoolItem.setMiners(objConfigPool.getInt("miners"));
+
+                        // Hashrate
+                        mPoolItem.setHr(Utils.convertStringToFloat(objConfigPool.getString("hashrate")) / 1000.0f);
+
+                        mPoolItem.setIsValid(true);
+                    } catch (Exception e) {
+                        mPoolItem.setIsValid(false);
+                    }
+                    finally {
+                        poolsAdapter.dataSetChanged();
+                    }
+                }
+                , PoolActivity::parseVolleyError);
+    }
     @Override
     protected void onBackgroundFetchData() {
         PrettyTime pTime = new PrettyTime();
@@ -44,7 +77,7 @@ public class CryptonoteNodejsPool extends ProviderAbstract {
 
             JSONObject joStats = new JSONObject(dataStatsNetwork);
             JSONObject joStatsConfig = joStats.getJSONObject("config");
-            //JSONObject joStatsLastBlock = joStats.getJSONObject("lastblock");
+            JSONObject joStatsLastBlock = joStats.has("lastblock") ? joStats.getJSONObject("lastblock") : null;
             JSONObject joStatsNetwork = joStats.getJSONObject("network");
             JSONObject joStatsPool = joStats.getJSONObject("pool");
 
@@ -54,17 +87,23 @@ public class CryptonoteNodejsPool extends ProviderAbstract {
             mBlockData.coin.denominationUnit = tryParseLong(joStatsConfig.optString("denominationUnit"), 1L);
 
             //mBlockData.pool.lastBlockHeight = joStatsPool.optString("height");
-            mBlockData.pool.difficulty = getReadableHashRateString(joStatsPool.optLong("totalDiff"));
             mBlockData.pool.lastBlockTime = pTime.format(new Date(joStatsPool.optLong("lastBlockFound")));
             //mBlockData.pool.lastRewardAmount = parseCurrency(joStatsLastBlock.optString("reward", "0"), mBlockData.coin.units, mBlockData.coin.denominationUnit, mBlockData.coin.symbol);
-            mBlockData.pool.hashrate = String.valueOf(tryParseLong(joStatsPool.optString("hashrate"),0L) / 1000L);
+            mBlockData.pool.hashrate = getReadableHashRateString(tryParseLong(joStatsPool.optString("hashrate"),0L));
             mBlockData.pool.blocks = joStatsPool.optString("roundHashes", "0");
+            mBlockData.pool.miners = joStatsPool.optString("miners", "0");
             mBlockData.pool.minPayout = parseCurrency(joStatsConfig.optString("minPaymentThreshold", "0"), mBlockData.coin.units, mBlockData.coin.denominationUnit, mBlockData.coin.symbol);
 
             mBlockData.network.lastBlockHeight = joStatsNetwork.optString("height");
-            mBlockData.network.difficulty = getReadableHashRateString(joStatsNetwork.optLong("difficulty"));
-            mBlockData.network.lastBlockTime = pTime.format(new Date(joStatsNetwork.optLong("timestamp") * 1000));
-            mBlockData.network.lastRewardAmount = parseCurrency(joStatsNetwork.optString("reward", "0"), mBlockData.coin.units, mBlockData.coin.denominationUnit, mBlockData.coin.symbol);
+            mBlockData.network.difficulty = getReadableDifficultyString(joStatsNetwork.optLong("difficulty"));
+
+            long lastBlock = (joStatsLastBlock != null ? joStatsLastBlock.optLong("timestamp") : joStatsNetwork.optLong("timestamp")) * 1000;
+            mBlockData.network.lastBlockTime = pTime.format(new Date(lastBlock));
+
+            String lastReward = joStatsLastBlock != null ? joStatsLastBlock.optString("reward", "0") : joStatsNetwork.optString("reward", "0");
+            mBlockData.network.lastRewardAmount = parseCurrency(lastReward, mBlockData.coin.units, mBlockData.coin.denominationUnit, mBlockData.coin.symbol);
+
+            mBlockData.network.hashrate = getReadableHashRateString(joStatsNetwork.optLong("difficulty") / 120L);
         } catch (JSONException e) {
             Log.i(LOG_TAG, "NETWORK\n" + e.toString());
             e.printStackTrace();
@@ -88,7 +127,7 @@ public class CryptonoteNodejsPool extends ProviderAbstract {
             String balance = parseCurrency(joStatsAddressStats.optString("balance", "0"), coin.units, coin.denominationUnit, coin.symbol);
             String paid = parseCurrency(joStatsAddressStats.optString("paid", "0"), coin.units, coin.denominationUnit, coin.symbol);
             String lastShare = pTime.format(new Date(joStatsAddressStats.optLong("lastShare") * 1000));
-            String blocks = String.valueOf(tryParseLong(joStatsAddressStats.optString("blocks"), 0L));
+            String blocks = String.valueOf(tryParseLong(joStatsAddressStats.optString("shares_good"), 0L));
 
             Log.i(LOG_TAG, "hashRate: " + hashRate);
 
@@ -96,7 +135,13 @@ public class CryptonoteNodejsPool extends ProviderAbstract {
             mBlockData.miner.balance = balance;
             mBlockData.miner.paid = paid;
             mBlockData.miner.lastShare = lastShare;
-            mBlockData.miner.blocks = blocks;
+            mBlockData.miner.shares = blocks;
+
+            // Payments
+            mBlockData.miner.payments.clear();
+
+            //TODO
+
         } catch (JSONException e) {
             Log.i(LOG_TAG, "ADDRESS\n" + e.toString());
             e.printStackTrace();
